Committer: Andrey Dmitrievich Vlasov <av@vc.localdomain>  Changes to be committed: 	new file:   surfaces/getMaksTel.m The function which creates new struct, MaksTel
	new file:   tracing/traceThroughMaksTel.m
the function which traces rays through new struct MaksTel

 Changes not staged for commit:
	modified:   initEnvio.m

Here I introduced a new parameter to initEnvio functions, windows. It's boolean parameter telling if we are on Windows.
if we are on windows, folder separator is \ (Linux it's /). Also, if we are in Windows, I disable opengl functionality because
Octave doesn't have it yet.

	modified:   tracing/refractThroughtQuad.m
Not significant changes (like whitespaces etc.)